Update lisp.tex
This edit was dismissed in class after being called, but I still believe it is required. Below is why:

All edits to the a_list happen through bind:
(t ; both names and values are not empty
(cons ; create new binding and prepend it to result of recursive call
(cons (car names) (car values)) ; new DOTTED-PAIR defines single binding
(bind (cdr names) (cdr values) a-list))))) ; recursive vall

notice that each pair in the a_list is dotted! Now look at the cdr usage examples at the end of the pdf:
(cdr ’(a.b)) ⇒ b
(cdr ’(a b)) ⇒ (b)

when running the PRIMITIVE cdr on a dotted pair, it returns the whole of the 2nd part in the dotted pair, not a list of it with nil.

If there is a flaw/misconception, I'd be happy to hear about it.